# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.WikiFileUploadResponse:
    properties:
      contentType:
        type: string
      filename:
        type: string
      id:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: Command-and-control core backend API.
  title: VibeC2 Core API
  version: "1.0"
paths:
...
  /wiki/files:
    post:
      consumes:
      - multipart/form-data
      description: Upload a non-image file to be attached to a wiki document. The
        original filename is preserved for downloads; bytes are stored as-is.
      parameters:
      - description: Owning document ID (UUID)
        in: formData
        name: documentId
        required: true
        type: string
      - description: File bytes (max 50MB by default)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.WikiFileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a wiki file attachment
      tags:
      - Wiki
  /wiki/files/{id}:
    get:
      description: Streams the file bytes. Caller must be a member of the file's operation.
        Pass ?preview=1 for inline rendering of safe formats (PDF, text, raster images).
      parameters:
      - description: File ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: If true, serve inline for safe MIME types
        in: query
        name: preview
        type: boolean
      produces:
      - application/octet-stream
      responses: {}
      security:
      - BearerAuth: []
      summary: Fetch a wiki file attachment
      tags:
      - Wiki
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token (e.g. "Bearer eyJhb...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"